dqms/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.2
python-dotenv==1.0.0
psycopg2-binary
PyJWT==2.7.0
